+++ csatar-plugins/csatar/Plugin.php
@@ -39,6 +39,7 @@ class Plugin extends PluginBase
         'PolloZen.SimpleGallery',
         'RainLab.Translate',
         'Rainlab.User',
+        'RainLab.Location',
         'Vdlp.TwoFactorAuthentication',
     ];
 
@@ -213,6 +214,7 @@ class Plugin extends PluginBase
             \Csatar\Csatar\Components\Breadcrumb::class => 'csatarBreadcrumb',
             \Csatar\Csatar\Components\TwoFactorAuthentication::class => 'twoFactorAuthentication',
             \Csatar\Csatar\Components\AccidentLogRecordList::class => 'accidentLogRecordList',
+            \Csatar\Csatar\Components\Partials::class => 'partials',
         ];
     }