+++ csatar-plugins/csatar/controllers/PermissionsMatrix.php
@@ -158,7 +158,7 @@ class PermissionsMatrix extends Controller
 
         if ($formData['action'] === 'copy') {
             $permissionsToCopy = MandatePermission::where('mandate_type_id', $formData['fromMandateType'])->get();
-            foreach ($permissionsToCopy as $permissionToCopy) { //dd($formData, $permissionToCopy);
+            foreach ($permissionsToCopy as $permissionToCopy) {
                 foreach ($formData['toMandateTypes'] as $toMandateTypeId) {
                     Db::table('csatar_csatar_mandates_permissions')
                         ->updateOrInsert(
@@ -228,7 +228,8 @@ class PermissionsMatrix extends Controller
                     //add permission for the model in general
                     $tempMandatePermissionsMap[] = [ 'mandate_type_id' => $mandateType->id, 'model' => $permissionBasedModel, 'field' => 'MODEL_GENERAL', 'own' => 0];
 
-                    if ($mandateType->organization_type_model_name != MandateType::MODEL_NAME_GUEST) {
+                    if ($mandateType->organization_type_model_name == MandateType::MODEL_NAME_SCOUT && $permissionBasedModel == MandateType::MODEL_NAME_SCOUT) {
+
                         //add permission for the model in general for own
                         $tempMandatePermissionsMap[] = [ 'mandate_type_id' => $mandateType->id, 'model' => $permissionBasedModel, 'field' => 'MODEL_GENERAL', 'own' => 1];
                     }
@@ -237,7 +238,7 @@ class PermissionsMatrix extends Controller
                     foreach ($fields as $field) {
                         $tempMandatePermissionsMap[] = [ 'mandate_type_id' => $mandateType->id, 'model' => $permissionBasedModel, 'field' => $field, 'own' => 0];
 
-                        if ($mandateType->organization_type_model_name != MandateType::MODEL_NAME_GUEST) {
+                        if ($mandateType->organization_type_model_name == MandateType::MODEL_NAME_SCOUT && $permissionBasedModel == MandateType::MODEL_NAME_SCOUT) {
                             $tempMandatePermissionsMap[] = ['mandate_type_id' => $mandateType->id, 'model' => $permissionBasedModel, 'field' => $field, 'own' => 1];
                         }
                     }