+++ csatar-plugins/csatar/models/Scout.php
@@ -19,6 +19,7 @@ use Model;
 use October\Rain\Database\Collection;
 use Session;
 use ValidationException;
+use Rainlab\Location\Models\Country;
 
 /**
  * Model
@@ -35,6 +36,13 @@ class Scout extends OrganizationBase
 
     protected static $relationLabels = null;
     public $active_mandates = [];
+
+    /**
+     * @var bool skipCacheRefresh
+     * If set to true, the cache will not be refreshed after save
+     * Usefull after bulk import, in such cases cache refresh should be done after all records are imported
+     */
+    public bool $skipCacheRefresh = false;
 
     /**
      * @var array The columns that should be searchable by ContentPageSearchProvider
@@ -60,6 +68,7 @@ class Scout extends OrganizationBase
         'nickname',
         'email',
         'phone',
+        'citizenship_country_id',
         'personal_identification_number',
         'gender',
         'is_active',
@@ -164,6 +173,7 @@ class Scout extends OrganizationBase
         'patrol' => 'nullable',
         'phone' => 'nullable|regex:(^[0-9+-.()]{10,}$)',
         'birthdate' => 'required',
+        'citizenship_country' => 'required',
         'legal_representative_phone' => 'nullable|regex:(^[0-9+-.()]{10,}$)',
         'personal_identification_number' => 'nullable',
         'mothers_phone' => 'nullable|regex:(^[0-9+-.()]{10,}$)',
@@ -229,17 +239,24 @@ class Scout extends OrganizationBase
                 return;
             }
 
-            $personalIdentificationNumberValidators = $this->getPersonalIdentificationNumberValidators();
+            // personal id number validations, for active scouts only
+            if ($this->is_active) {
+                $personalIdentificationNumberValidators = $this->getPersonalIdentificationNumberValidators();
 
-            if (!empty($personalIdentificationNumberValidators)) {
-                $this->rules['personal_identification_number'] .= '|' . implode('|', $personalIdentificationNumberValidators);
-            }
+                if (in_array('cnp', $personalIdentificationNumberValidators) && $this->shouldNotValidateCnp()) {
+                    unset($personalIdentificationNumberValidators[array_search('cnp', $personalIdentificationNumberValidators)]);
+                }
 
-            if (in_array('cnp', $personalIdentificationNumberValidators)
-                && !empty($this->personal_identification_number)
-                && (new DateTime($this->birthdate))->format('Y-m-d') != $this->getBirthDateFromCNP($this->personal_identification_number)
-            ) {
-                throw new \ValidationException(['birthdate' => Lang::get('csatar.csatar::lang.plugin.admin.scout.validationExceptions.personalIdentificationNumberBirthdateMismatch')]);
+                if (!empty($personalIdentificationNumberValidators)) {
+                    $this->rules['personal_identification_number'] .= '|' . implode('|', $personalIdentificationNumberValidators);
+                }
+
+                if (in_array('cnp', $personalIdentificationNumberValidators)
+                    && !empty($this->personal_identification_number)
+                    && (new DateTime($this->birthdate))->format('Y-m-d') != $this->getBirthDateFromCNP($this->personal_identification_number)
+                ) {
+                    throw new \ValidationException(['birthdate' => Lang::get('csatar.csatar::lang.plugin.admin.scout.validationExceptions.personalIdentificationNumberBirthdateMismatch')]);
+                }
             }
 
             // if the selected troop does not belong to the selected team, then throw and exception
@@ -290,18 +307,40 @@ class Scout extends OrganizationBase
             }
         }
 
-        if (empty($this->original)) {
+        if (!$this->skipCacheRefresh) {
+            $this->updateCache();
+        }
+
+
+    }
+
+    public function afterDelete() {
+        if ($this->skipCacheRefresh) {
+            return;
+        }
+        if (!empty($this->team_id)) {
+            StructureTree::updateTeamTree($this->team_id);
+        }
+    }
+
+    public function updateCache(): void
+    {
+        if ($this->wasRecentlyCreated && $this->is_active == Status::ACTIVE) {
+            StructureTree::updateTeamTree($this->team_id);
+        }
+
+        if (empty($this->original) ) {
             return;
         }
 
-        if (isset($this->original['is_active']) && $this->original['is_active'] != $this->is_active) {
+        if (($this->getOriginalValue('is_active') != $this->is_active) || $this->deleted_at != null) {
             StructureTree::updateTeamTree($this->team_id);
         }
 
         if (
-            (isset($this->original['team_id']) && $this->original['team_id'] != $this->team_id)
-            || (isset($this->original['troop_id']) && $this->original['troop_id'] != $this->troop_id)
-            || (isset($this->original['patrol_id']) && $this->original['patrol_id'] != $this->patrol_id)
+            ($this->getOriginalValue('team_id') != $this->team_id)
+            || ($this->getOriginalValue('troop_id') != $this->troop_id)
+            || ($this->getOriginalValue('patrol_id') != $this->patrol_id)
         )
         {
             StructureTree::updateTeamTree($this->team_id);
@@ -311,10 +350,10 @@ class Scout extends OrganizationBase
         }
 
         if (
-            (isset($this->original['family_name']) && $this->original['family_name'] != $this->family_name)
-            || (isset($this->original['given_name']) && $this->original['given_name'] != $this->given_name)
-            || (isset($this->original['ecset_code']) && $this->original['ecset_code'] != $this->given_name)
-            || (isset($this->original['legal_relationship_id']) && $this->original['legal_relationship_id'] != $this->legal_relationship_id)
+            ($this->getOriginalValue('family_name') != $this->family_name)
+            || ($this->getOriginalValue('given_name') != $this->given_name)
+            || ($this->getOriginalValue('ecset_code') != $this->ecset_code)
+            || ($this->getOriginalValue('legal_relationship_id') != $this->legal_relationship_id)
         )
         {
             $structureTree = Cache::pull('structureTree');
@@ -372,14 +411,16 @@ class Scout extends OrganizationBase
             ->where(function ($query) {
                 $query->whereNull('end_date')
                   ->orWhere('end_date', '>=', date('Y-m-d H:i'));
-            })->get();
+            })
+            ->with('mandate_type')
+            ->get();
 
         // first add the team mandates in the mandates list
         foreach ($mandates as $key => $value) {
             if ($value->mandate_model_type == '\Csatar\Csatar\Models\Team') {
                 array_push($this->active_mandates, [
                     'title' => '',
-                    'value' => isset(MandateType::find($value->mandate_type_id)->name) ? MandateType::find($value->mandate_type_id)->name : '',
+                    'value' => $value->mandate_type->name ?? '',
                 ]);
             }
         }
@@ -389,7 +430,7 @@ class Scout extends OrganizationBase
             if ($value->mandate_model_type != '\Csatar\Csatar\Models\Team') {
                 array_push($this->active_mandates, [
                     'title' => $value->mandate_model_name,
-                    'value' => isset(MandateType::find($value->mandate_type_id)->name) ? MandateType::find($value->mandate_type_id)->name : '',
+                    'value' => $value->mandate_type->name ?? ''
                 ]);
             }
         }
@@ -440,6 +481,7 @@ class Scout extends OrganizationBase
         }
 
         if (isset($fields->personal_identification_number)
+            && !$this->shouldNotValidateCnp()
             && !empty($fields->personal_identification_number->value)
             && in_array('cnp', $this->getPersonalIdentificationNumberValidators())
             && ((isset($this->original['personal_identification_number']) && $this->original['personal_identification_number'] != $fields->personal_identification_number->value) || empty($this->original))
@@ -457,6 +499,10 @@ class Scout extends OrganizationBase
 
         if (isset($fields->address_street)) {
             $this->setAddressStreetOptions($fields->address_street);
+        }
+
+        if (isset($fields->citizenship_country) && empty($fields->citizenship_country->value)) {
+            $fields->citizenship_country->value = Country::where('code', 'RO')->first()->id ?? null;
         }
     }
 
@@ -480,6 +526,11 @@ class Scout extends OrganizationBase
         ],
         'troop' => '\Csatar\Csatar\Models\Troop',
         'patrol' => '\Csatar\Csatar\Models\Patrol',
+        'citizenship_country' => [
+            '\Rainlab\Location\Models\Country',
+            'label' => 'csatar.csatar::lang.plugin.admin.scout.citizenship_country',
+            'key' => 'citizenship_country_id',
+            ],
     ];
 
     public $belongsToMany = [
@@ -580,6 +631,37 @@ class Scout extends OrganizationBase
         'registration_form' => 'System\Models\File',
     ];
 
+    public static function getEagerLoadSettings(string $useCase = null): array
+    {
+        $eagerLoadSettings = parent::getEagerLoadSettings($useCase);
+        if ($useCase === 'formBuilder') {
+            // Important to extend the eager load settings, not to overwrite them!
+            $eagerLoadSettings = array_merge_recursive($eagerLoadSettings, [
+                'allergies', 'chronic_illnesses', 'food_sensitivities', 'promises', 'tests', 'special_tests', 'professional_qualifications', 'special_qualifications', 'leadership_qualifications', 'training_qualifications', 'team', 'team.district', 'team.district.association', 'troop', 'patrol'
+            ]);
+        }
+        return $eagerLoadSettings;
+    }
+
+    public function getTeamOptions() {
+        $teams = Team::forDropdown()->get();
+        $teamOptions = [];
+        foreach ($teams as $team) {
+            $teamOptions[$team->id] = $team->extended_name_with_association;
+        }
+        return $teamOptions;
+    }
+
+    public function getCitizenshipCountryOptions() {
+        $countries = Country::all();
+        $countryOptions = [];
+        foreach ($countries as $country) {
+            $countryOptions[$country->id] = $country->getAttributeTranslated('name', 'hu');
+            // this is a hardcoded language setting, will should be solved with task CS-521
+        }
+        return $countryOptions;
+    }
+
     public function beforeCreate()
     {
         $this->ecset_code = isset($this->ecset_code) && !empty($this->ecset_code) ? $this->ecset_code : strtoupper($this->generateEcsetCode());
@@ -611,7 +693,7 @@ class Scout extends OrganizationBase
         $this->patrol_id = $this->patrol_id != 0 ? $this->patrol_id : null;
     }
 
-    function beforeDelete()
+    public function beforeDelete()
     {
         $now = new DateTime();
         $mandates = Mandate::where('scout_id', $this->id)->get();
@@ -622,6 +704,10 @@ class Scout extends OrganizationBase
                 return false;
             }
         }
+    }
+
+    private function shouldNotValidateCnp() {
+        return empty($this->citizenship_country_id) || (!empty($this->citizenship_country_id) && $this->citizenship_country_id != (new CnpValidator())->getRomaniaCountryId());
     }
 
     public function setAllMandatesToExpiredInOrganization(OrganizationBase $organization): void
@@ -965,17 +1051,15 @@ class Scout extends OrganizationBase
     }
 
     public function getMandatesForOrganization(PermissionBasedAccess $organization, bool $withInactive = false) {
-        return $this->mandates()
+        return $this->mandates
             ->where('mandate_model_type', $organization->getModelName())
             ->where('mandate_model_id', $organization->id)
-            ->when(!$withInactive, function ($query){
-                $query->where('start_date', '<=', date('Y-m-d H:i'))
-                      ->where(function ($query) {
-                    $query->whereNull('end_date')
-                          ->orWhere('end_date', '>=', date('Y-m-d H:i'));
-                });
-            })
-            ->get();
+            ->when(!$withInactive, function ($collection){
+                return $collection->where('start_date', '<=', date('Y-m-d H:i'))
+                    ->filter(function ($item) {
+                        return $item->end_date === null || $item->end_date >= date('Y-m-d H:i');
+                    });
+            });
     }
 
     public function saveMandateTypeIdsForEveryAssociationToSession(){
@@ -1004,6 +1088,8 @@ class Scout extends OrganizationBase
         if (empty($model)) {
             return;
         }
+
+        $this->load('mandates', 'mandates.mandate_type');
 
         $isOwn = false;
         if(Auth::user() && !empty(Auth::user()->scout)){
@@ -1148,8 +1234,7 @@ class Scout extends OrganizationBase
 
     public function setAddressCountyOptions(&$field)
     {
-        $savedCountyArray = Scout::where('id', $this->id)->select('address_county')->first();
-        $savedCounty = $savedCountyArray['address_county'] ?? null;
+        $savedCounty = $this->original['address_county'] ?? null;
         $array = [];
         if ($this->address_zipcode != null) {
             $array = Locations::where('country', '=', $this->address_country)->where('code', '=', $this->address_zipcode)->lists('county', 'county');
@@ -1170,8 +1255,7 @@ class Scout extends OrganizationBase
 
     public function setAddressLocationOptions(&$field)
     {
-        $savedLocationArray = Scout::where('id', $this->id)->select('address_location')->first();
-        $savedLocation = $savedLocationArray['address_location'] ?? null;
+        $savedLocation = $this->original['address_location'] ?? null;
         $array = [];
 
         if ($this->address_zipcode != null) {
@@ -1193,8 +1277,7 @@ class Scout extends OrganizationBase
 
     public function setAddressStreetOptions(&$field)
     {
-        $savedStreetArray = Scout::where('id', $this->id)->select('address_street')->first();
-        $savedStreet = $savedStreetArray['address_street'] ?? null;
+        $savedStreet = $this->original['address_street'] ?? null;
         $array = [];
 
         if ($this->address_zipcode != null) {
@@ -1223,8 +1306,7 @@ class Scout extends OrganizationBase
     public function getAddressCountryAttribute()
     {
         $savedCountry = array_get($this->attributes, 'address_country');
-        $teamId = array_get($this->attributes, 'team_id');
-        $team = Team::find($teamId);
+        $team = $this->team ?? Team::find($this->team_id);
 
         if (empty($team)) {
             return null;