+++ csatar-plugins/csatar/models/Troop.php
@@ -23,6 +23,8 @@ class Troop extends OrganizationBase
 
     protected $appends = ['extended_name'];
 
+    public $customAttributes = ['active_members_count'];
+
     /**
      * @var array Validation rules
      */
@@ -101,6 +103,44 @@ class Troop extends OrganizationBase
         'logo' => 'System\Models\File'
     ];
 
+    public static function getEagerLoadSettings(string $useCase = null): array
+    {
+        $eagerLoadSettings = parent::getEagerLoadSettings($useCase);
+        if ($useCase === 'formBuilder') {
+            // Important to extend the eager load settings, not to overwrite them!
+            $eagerLoadSettings['mandates.mandate_troop'] = function($query) {
+                return $query->select(
+                    'csatar_csatar_troops.id',
+                    'csatar_csatar_troops.team_id'
+                );
+            };
+            $eagerLoadSettings['mandates.mandate_troop.team'] = function($query) {
+                return $query->select(
+                    'csatar_csatar_teams.id',
+                    'csatar_csatar_teams.name',
+                    'csatar_csatar_teams.team_number',
+                    'csatar_csatar_teams.district_id'
+                );
+            };
+            $eagerLoadSettings = array_merge_recursive($eagerLoadSettings, [
+                'team', 'team.district', 'team.district.association',
+            ]);
+        }
+        if ($useCase == 'inactiveMandatesTroop') {
+            $eagerLoadSettings = [
+                'mandatesInactive.mandate_troop.team' => function($query) {
+                    return $query->select(
+                        'csatar_csatar_teams.id',
+                        'csatar_csatar_teams.name',
+                        'csatar_csatar_teams.team_number'
+                    )->withTrashed();
+                },
+            ];
+            $eagerLoadSettings = array_merge($eagerLoadSettings, parent::getEagerLoadSettings('inactiveMandates'));
+        }
+        return $eagerLoadSettings;
+    }
+
     public function beforeSave()
     {
         $filterWords = explode(',', Lang::get('csatar.csatar::lang.plugin.admin.troop.filterOrganizationUnitNameForWords'));
@@ -125,22 +165,31 @@ class Troop extends OrganizationBase
             Mandate::setAllMandatesExpiredInOrganization($this);
         }
 
+        $this->updateCache();
+    }
+
+    public function updateCache(): void
+    {
+        if ($this->wasRecentlyCreated && $this->status == Status::ACTIVE) {
+            StructureTree::updateAssociationTree($this->association_id);
+        }
+
         if (empty($this->original)) {
             return;
         }
 
-        if (isset($this->original['status']) && $this->original['status'] != $this->status) {
+        if ($this->getOriginalValue('status') != $this->status) {
             StructureTree::updateTeamTree($this->team_id);
         }
 
-        if (isset($this->original['team_id']) && $this->original['team_id'] != $this->district_id) {
+        if ($this->getOriginalValue('team_id') != $this->district_id) {
             StructureTree::updateTeamTree($this->team_id);
             if (!empty($this->original['team_id'])) {
                 StructureTree::updateTeamTree($this->original['team_id']);
             }
         }
 
-        if (isset($this->original['name']) && $this->original['name'] != $this->name) {
+        if ($this->getOriginalValue('name') != $this->name) {
             $structureTree = Cache::pull('structureTree');
             if (empty($structureTree)) {
                 StructureTree::getStructureTree();
@@ -252,5 +301,9 @@ class Troop extends OrganizationBase
     public function scopeActive($query)
     {
         return $query->where('status', Status::ACTIVE);
+    }
+
+    public function getActiveMembersCountAttribute() {
+        return StructureTree::getTroopScoutsCount($this->id);
     }
 }