+++ csatar-plugins/knowledgerepository/models/Headcount.php
@@ -5,7 +5,7 @@ use Model;
 /**
  * Model
  */
-class HeadCount extends Model
+class Headcount extends Model
 {
     use \October\Rain\Database\Traits\Validation;
     use \October\Rain\Database\Traits\Sortable;